fix circular imports in RtcpPacketConverter
