+++ packages/rtp/src/rtcp/sdes.ts
@@ -1,6 +1,5 @@
 import { bufferWriter } from "../../../common/src";
-import type { RtcpHeader } from "./header";
-import { RtcpPacketConverter } from "./rtcp";
+import { RtcpHeader } from "./header";
 
 export class RtcpSourceDescriptionPacket {
   static readonly type = 202;
@@ -21,7 +20,7 @@ export class RtcpSourceDescriptionPacket {
     let payload = Buffer.concat(this.chunks.map((chunk) => chunk.serialize()));
     while (payload.length % 4)
       payload = Buffer.concat([payload, Buffer.from([0])]);
-    return RtcpPacketConverter.serialize(
+    return RtcpHeader.serialize(
       this.type,
       this.chunks.length,
       payload,
