+++ packages/rtp/tests/rtcp/rtcp.test.ts
@@ -1,4 +1,4 @@
-import { PictureLossIndication, RtcpPayloadSpecificFeedback } from "../../src";
+import { PictureLossIndication, RtcpHeader, RtcpPayloadSpecificFeedback } from "../../src";
 import { RtcpRrPacket } from "../../src/rtcp/rr";
 import { RtcpPacketConverter } from "../../src/rtcp/rtcp";
 import { RtcpSrPacket } from "../../src/rtcp/sr";
@@ -57,7 +57,7 @@ describe("rtcp/rtcp", () => {
   });
 
   test("should fail", () => {
-    const rtcp = RtcpPacketConverter.serialize(
+    const rtcp = RtcpHeader.serialize(
       RtcpPayloadSpecificFeedback.type,
       PictureLossIndication.count,
       Buffer.from([0]),
